Datattable
adição de toda a secção "Projetos", adição de mecanismo de segurança para rotas destroy/delete adição de icons nas defenições adiçao do end date e mecanismos logicos

+++ RoboticsCodeRaulWeb/routes/web.php
@@ -11,7 +11,7 @@ Route::redirect('/', '/home');
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
 //Index Back
-Route::get('/dashboard', [App\Http\Controllers\BackController::class, 'index'])->name('dashboard');
+Route::get('/dashboard', [App\Http\Controllers\BackController::class, 'index'])->name('dashboard')->middleware('auth');
 
 //Auth
 Route::get('/register', [App\Http\Controllers\Auth\RegisterController::class, 'showRegistrationForm'])->name('register');
@@ -27,28 +27,30 @@ Route::get('/sobre-nos', [App\Http\Controllers\FrontController::class, 'about_us
 Route::get('/concursos', [App\Http\Controllers\FrontController::class, 'contests'])->name('contests');
 
 //Backoffice
-Route::get('/acessos', [App\Http\Controllers\AccessController::class, 'index'])->name('access');
+Route::get('/acessos', [App\Http\Controllers\AccessController::class, 'index'])->name('access')->middleware('auth');
 
-Route::get('/conta', [App\Http\Controllers\AccountController::class, 'index'])->name('account');
+Route::get('/conta', [App\Http\Controllers\AccountController::class, 'index'])->name('account')->middleware('auth');
 
 // web.php
-Route::get('/membros', [App\Http\Controllers\MemberController::class, 'index'])->name('members');
-Route::get('/membros/criar', [App\Http\Controllers\MemberController::class, 'create'])->name('members.create');
-Route::post('/membros', [App\Http\Controllers\MemberController::class, 'store'])->name('members.store');
-Route::get('/membros/{member}', [App\Http\Controllers\MemberController::class, 'show'])->name('members.show');
-Route::get('/membros/{member}/editar', [App\Http\Controllers\MemberController::class, 'edit'])->name('members.edit');
-Route::put('/membros/{member}', [App\Http\Controllers\MemberController::class, 'update'])->name('members.update');
-Route::delete('/membros/{member}', [App\Http\Controllers\MemberController::class, 'destroy'])->name('members.destroy');
+Route::get('/membros', [App\Http\Controllers\MemberController::class, 'index'])->name('members')->middleware('auth');
+Route::get('/membros/criar', [App\Http\Controllers\MemberController::class, 'create'])->name('members.create')->middleware('auth');
+Route::post('/membros', [App\Http\Controllers\MemberController::class, 'store'])->name('members.store')->middleware('auth');
+Route::get('/membros/{member}/visualizacao', [App\Http\Controllers\MemberController::class, 'show'])->name('members.show')->middleware('auth');
+Route::get('/membros/{member}/editar', [App\Http\Controllers\MemberController::class, 'edit'])->name('members.edit')->middleware('auth');
+Route::put('/membros/{member}', [App\Http\Controllers\MemberController::class, 'update'])->name('members.update')->middleware('auth');
+Route::delete('/membros/{member}', [App\Http\Controllers\MemberController::class, 'destroy'])->name('members.destroy')->middleware('auth');
 
-Route::get('noticias', [App\Http\Controllers\NewsController::class, 'index'])->name('news');
-Route::get('noticias/criar', [App\Http\Controllers\NewsController::class, 'create'])->name('news.create');
-Route::post('noticias', [App\Http\Controllers\NewsController::class, 'store'])->name('news.store');
-Route::get('noticias/{news}', [App\Http\Controllers\NewsController::class, 'show'])->name('news.show');
+Route::get('noticias', [App\Http\Controllers\NewsController::class, 'index'])->name('news')->middleware('auth');
+Route::get('noticias/criar', [App\Http\Controllers\NewsController::class, 'create'])->name('news.create')->middleware('auth');
+Route::post('noticias', [App\Http\Controllers\NewsController::class, 'store'])->name('news.store')->middleware('auth');
+Route::get('noticias/{news}', [App\Http\Controllers\NewsController::class, 'show'])->name('news.show')->middleware('auth');
 
-Route::get('/projetos', [App\Http\Controllers\ProjectController::class, 'index'])->name('projects');
-Route::get('/projetos/criar', [App\Http\Controllers\ProjectController::class, 'create'])->name('projects.create');
-Route::post('/projetos', [App\Http\Controllers\ProjectController::class, 'store'])->name('projects.store');
-Route::get('/projetos/{projects}', [App\Http\Controllers\ProjectController::class, 'show'])->name('projects.show');
-
+Route::get('/projetos', [App\Http\Controllers\ProjectController::class, 'index'])->name('projects')->middleware('auth');
+Route::get('/projetos/criar', [App\Http\Controllers\ProjectController::class, 'create'])->name('projects.create')->middleware('auth');
+Route::post('/projetos', [App\Http\Controllers\ProjectController::class, 'store'])->name('projects.store')->middleware('auth');
+Route::get('/projetos/{project}/visualizacao', [App\Http\Controllers\ProjectController::class, 'show'])->name('projects.show')->middleware('auth');
+Route::get('/projetos/{project}/editar', [App\Http\Controllers\ProjectController::class, 'edit'])->name('projects.edit')->middleware('auth');
+Route::put('/projetos/{project}', [App\Http\Controllers\ProjectController::class, 'update'])->name('projects.update')->middleware('auth');
+Route::delete('/projetos/{project}', [App\Http\Controllers\ProjectController::class, 'destroy'])->name('projects.destroy')->middleware('auth');
 
 Auth::routes();
